@@ -1,14 +1,14 @@
 {
   "name": "ngx-gauge",
-  "version": "1.0.0-beta.7",
+  "version": "1.0.0-beta.11",
   "description": "A highly customizable Gauge Component for Angular 4+ apps and dashboards",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ashish-chopra/ngx-gauge.git"
   },
   "peerDependencies": {
-    "@angular/common": "^7.1.0",
-    "@angular/core": "^7.1.0"
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0"
   },
   "keywords": [
     "angular",
