@@ -1,14 +1,21 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-export class AppComponent {
+export class AppComponent implements OnInit {
   title = 'demo';
   currentValue: number = 0;
+  time = 700;
   changeValue() {
-    this.currentValue = parseFloat((Math.random() * 100).toFixed(2));
+    //nothing
+  }
+
+  ngOnInit() {
+    setInterval(() => {
+      this.currentValue = Math.floor(Math.random() * 100);
+    }, this.time);
   }
 }
